PROACTIVE-571: Added GCMApplication.waitReady(long timeout)
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@10995 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Core/org/objectweb/proactive/gcmdeployment/GCMVirtualNode.java
@@ -118,7 +118,7 @@ public interface GCMVirtualNode {
      *             If the timeout is reached
      */
     // @snippet-start GCMVirtualNode_waitReady_timeout
-    public void waitReady(int timeout) throws ProActiveTimeoutException;
+    public void waitReady(long timeout) throws ProActiveTimeoutException;
 
     // @snippet-end GCMVirtualNode_waitReady_timeout
